pyproject.toml
[project]
name = "dynamodb-pract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "boto3",
    "matplotlib",
    "seaborn",
    "python-dotenv",
    "tabulate",
]
